Close gh-20 Support for Google Chrome
The downside is that we can no lnger support the API version 1.0 which means that old browsers or Chrome using an old version of the extension is FUBAR - all that because FIDO stupidly changed the method signature WITHOUT providing a way for developers to detect which bloody version of the API is available!!! What a short-sighted decision. Boo!